refactor(fuse): extract pure convert helpers into convert.rs (#217)
